devman: detect acpi irq
git-svn-id: svn://kolibrios.org@1633 a494cfbc-eb01-0410-851d-a64ba20cac60

+++ drivers/devman/pci/access.c
@@ -77,7 +77,7 @@ int pci_bus_read_config_word (struct pci_bus *bus, u32 devfn,
 
 
 int pci_bus_read_config_dword (struct pci_bus *bus, u32 devfn,
-                              int pos, u16 *value)
+                               int pos, u32 *value)
 {
     if ( pos & 3)
         return PCIBIOS_BAD_REGISTER_NUMBER;
